- Annotations - Moving inner class to specific file
