+++ onecalendar/src/main/java/com/darwindeveloper/onecalendar/domain/OnSwipeListener.java
@@ -1,0 +1,7 @@
+package com.darwindeveloper.onecalendar.domain;
+
+public interface OnSwipeListener {
+    void rightSwipe();
+
+    void leftSwipe();
+}
